Add JSON output support to dmstats for list and report subcommands
Support "--reportformat basic|json" for the dmstats command.

In the case where --count is specified with the JSON reportformat, the
output will be <count> JSON documents.

Example output of a single iteration in JSON:

"dmstats": [{
      "stats_name": "mirrorvg-mirrorlv",
      "group_id": "-",
      "region_id": "0",
      "obj_type": "area",
      "area_id": "0",
      "area_start": "0",
      "area_len": "1.00g",
      "reads_merged_per_sec": "0.00",
      "writes_merged_per_sec": "0.00",
      "reads_per_sec": "0.00",
      "writes_per_sec": "0.00",
      "read_size_per_sec": "0",
      "write_size_per_sec": "0",
      "avg_request_size": "0",
      "queue_size": "0.00",
      "util": "0.00",
      "await": "0.00",
      "read_await": "0.00",
      "write_await": "0.00"
   }]
}

The above JSON has been reformatted due to line length, but the content
is from dmstats.

Signed-off-by: Todd Gill <tgill@redhat.com>